Bluetooth: UBSan fix for HAP
Addressed UBSan errors reported on hearing aid

+++ bta/hearing_aid/hearing_aid.cc
@@ -1183,9 +1183,9 @@ class HearingAidImpl : public HearingAid {
         sample += 2;
         int16_t right = (int16_t)((*(sample + 1) << 8) + *sample) >> 1;
 
-        uint16_t mono_data = (int16_t)(((uint32_t)left + (uint32_t)right) >> 1);
-        chan_left.push_back(mono_data);
-        chan_right.push_back(mono_data);
+        int32_t mono_data = (int32_t)((left + right) >> 1);
+        chan_left.push_back(uint16_t(mono_data));
+        chan_right.push_back(uint16_t(mono_data));
       }
     } else {
       for (int i = 0; i < num_samples; i++) {
